dev 개발 완료 (#74)
* chore: init

* Update .gitignore

* feat: User 관련 Entity, Dto 추가

* docs: 자바 버전 17->21로 변경

* feat: JWT 코드 추가

* feat: 로그인 패키지 구조 추가

* feat: 회원가입 추가

* feat: 회원가입 - userDto, dto, entity 변환 함수 추가

* FEAT : Add Space Entity, Dto (closes #7)

* feat: 로그인 요청 및 응답 구현 DTO

* feat: JWT 기반 로그인 기능 구현

* [FEAT] Add space entity/dto (Comment 반영) (closes #7)

* FIX: currentMemberCount/maxMemberCount 변수명 수정 (closes #7)

* refactor: 회원가입 관련 코드 자잘한 수정

* feat: 이메일 중복 검사 기능 추가

* FEAT: Add RefreshToken, RefreshTokenRepository (#11)

* feat: refreshToken 로직 추가

* feat: room entity, repository 구현

* feat: dto 작성 및 entity, repository 수정

* feat: memberProfilecard/answer entity 추가

* refactor: 커스텀 필드 컬럼명 및 repository 메서드명 수정

* refactor: 커스텀 필드 타입 명칭을 OptionType으로 통일

* refactor: memberprofilecard 경로수정

* refactor: ManyToOne 수정

* Issue#20/refactor apiresponse,user,auth controller (#21)

* refactor: apiResponse 내용으로 수정

* refactor: user, login DTO response, request 패키지 생성 및 변경

* feat: 채팅 기능 구현 및 jwt인증처리까지 (#22)

* feat: 채팅 기능 구현 및 jwt인증처리까지

* feat: UserRole 이니셜라이저 추가

* refactor: 컨트롤러/서비스 apiResponse 리팩토링 및 ChatController권한 검사 추가, entity, dto 내 필요없는 어노테이션 삭제

* [Feat] 룸 CRUD 및 권한 검증 Service 구현 (#23)

* feat: 룸 CRUD 및 권한 검증 Service 구현

* feat: Room Controller 구현

* fix: room dto, controller 수정

* fix: 룸 조회 및 수정 오류 해결

* Isuue#15/space crud (#25)

* FEAT : Space CRUD + TAG

* TEST : Test 용 html, SecurityConfig 허용

* FIX : 테스트 페이지 직관적으로 변경, ServicImpl 로직 오류 수정

---------


* Issue#26/tag search (#28)

* FEAT : Space CRUD + TAG

* TEST : Test 용 html, SecurityConfig 허용

* FIX : 테스트 페이지 직관적으로 변경, ServicImpl 로직 오류 수정

* FEAT : Find Tags in Room, Test 추가 (#26)

* feat: 로그아웃 기능 구현 (#29)

* Issue#24/member profile card crud (#35)

* feat: MemberProfileCard CRUD

* refactor: 테스트html 삭제

* feat: 멤프카 중복검사, 및 필수 항목 검사 추가

* feat: update 항목에도 필수 항목 검사 추가

* refactor: /filter post로 변경

* feat: 필수 항목 검사 null도 추가

* feat: 필수 항목인데 답변이 없거나 빈칸인 경우 예외처리

* Issue#24/member profile card crud (#36)

* feat: MemberProfileCard CRUD

* refactor: 테스트html 삭제

* feat: 멤프카 중복검사, 및 필수 항목 검사 추가

* feat: update 항목에도 필수 항목 검사 추가

* refactor: /filter post로 변경

* feat: 필수 항목 검사 null도 추가

* feat: 필수 항목인데 답변이 없거나 빈칸인 경우 예외처리

* feat: 커스텀필드 추가/수정 시 오류 3가지 사항 추가

* [Feat] 룸-스페이스 권한 검증 추가 (#37)

* feat: 룸 삭제 시 소속 스페이스 삭제

* feat: 스페이스 룸 권한 검증 추가

* Issue#30 add member to space (#38)

* FEAT : Space CRUD + TAG

* TEST : Test 용 html, SecurityConfig 허용

* FIX : 테스트 페이지 직관적으로 변경, ServicImpl 로직 오류 수정

* FEAT : Find Tags in Room, Test 추가 (#26)

* FEAT : Space, SpaceMember 일부 메서드 수정, 스페이스 초대/권유, 신청 목록 조회 등 구현 (#30)

* FIX : Debug, testCode 추가, LazyInitializationException 수정 (#30)

* REFACTOR : 검증 로직 추가, 가독성 문제 개선, 컨트롤러 메서드 분리 (#30)

* FIX : Repository 쿼리 수정 (#30)

* FIX : RoomID 경로 검증 추가 (#30)

* Feat : Controller APi URL 변경 - /invites 로 통일 (#30)

* feat: 전역 및 도메인별 예외처리 (#43)

* feat: 전역 및 도메인별 예외처리

* docs: 예외처리 작성 가이드 추가

* FIX : SpaceMember 버그 수정, InviteSpace 수정, #FEAT : Space 위임, 멤버 Space 탈퇴 (#44)

* Issue#45/improve space (#46)

* FIX : SpaceMember 버그 수정, InviteSpace 수정, #FEAT : Space 위임, 멤버 Space 탈퇴

* FIX : bug 수정 (#39)

* FIX : 로직 최종 수정, TEST 코드 변경(#39)

* FIX : 수정 pr 불가로 재커밋 (#45)

* Issue#33/feat chat,member profile card j unit (#47)

* chore: Jacoco 설정

* chore: settings.gradle내용 삭제

* chore: 자잘한거 삭제

* [Feat] 룸 프로필 카드 커스텀 필드 추가/삭제 구현 (#42)

* feat: 룸 프로필 카드 커스텀 필드 추가/삭제 구현

* feat: 룸 커스텀 필드 일괄 수정 기능 구현

* refactor: 멤프카update로직 변경

기존: 전체 바꾸면 싹다 삭제하고 새로 추가, 안바꾸면 있는 값만 업데이트

현재: dto로 들어온 필드값 확인 ->
1.멤프카에 이미 존재하는 필드면 값만 업데이트
2. 멤프카에 있지만 dto엔 없는 필드는 삭제
3. 멤프카에 없지만 dto에 있는 필드는 새로 추가

* refactor: 룸 커스텀 필드를 추가/삭제 방식으로 변경

---------


* feat: 스페이스채팅 (#51)

* feat: 토큰에 유저네임추가, 방 목록 조회 추가 (#55)

* [Feat] 룸 및 커스텀 필드 서비스 단위 테스트 추가 (#58)

* feat: 룸 서비스 테스트 구조 추가

* test: 룸 및 커스텀 필드 서비스 단위 테스트 추가

* feat, refactor : chat, memberProfileCard JUnit Test + memberProfileCard, inviteSpace Refactor (#64)

* Issue#52/test front (#66)

* FEAT : ADD FrontCode Test 1 ( #52)

* DOCS: 버그리스트 md파일 추가

* fix: 스페이스채팅 입장불가 수정

쿼리에서 sm.user.id가 :userId랑 같은지 비교하는걸 빼먹어서 오류가 났었네요

* IMPROVE : narangnorang.html 수정, FEAT : spaceMemberList 조회 기능 추가
(#52)

* FIX : UI 수정

* FIX : Tag_Update_Bug(#56) (#57)

* REFACTOR : Space, inviteSpace refactoring (#59) (#60)

* feat: 다중선택 가능하게 수정

* FIX : 디자인 및 일부 버그 수정 (#52)

* fix: 프로필 카드 필수 항목 표시 및 입력 검증 추가

* fix: 룸 전환 시 캐시 및 채팅 상태 초기화

* FIX : 스페이스 삭제 에러

* fix: 필수 커스텀 필드 미응답 시 프로필 수정 유도

* fix: 모달 실행 시 채팅 내역 초기화 문제 수정

* FEAT : USerDetails 조회, User 탈퇴, User 업데이트 추가 (#65)

* FEAT : USerDetails 조회, User 탈퇴, User 업데이트 추가

* merge 방지용

* FIX : ExceptionHandler 추가

* FIX : 오타 수정

---------


* Issue#67/front UI (#69)

* FIX : UI 수정

* feat:

* feat: 멤버 필터링 프론트 추가

* fix: 룸 생성 후 멤버 프로필 카드 작성 화면으로 이동

* FEAT : LocalDateTime -> TIME 으로 변경

* feat: 스페이스 관리에 태그 검색 토글 추가

* feat: 스페이스 정보 수정 커스텀필드 가능하게 변경

* feat: 룸 및 스페이스 멤버 태그 추가

* fix: 모달 리다이렉팅 문제 수정

* feat: 스페이스 관리창에서 위임 이상하게 되던거 수정

* fix: 모달 동작 시 채팅 화면 유지

* feat: 스페이스 참여중/미참여/오너 표시 추가 및 필수항목 표시 추가

* feat: 채팅 순서 및 날짜 추가, 스페이스 열고접기, 필터 위로

* feat: 다른 사람이 보낸 채팅 말풍선 길이 조절

* refactor: 스페이스 탈퇴누르고 확인 누른다음에 채팅 제대로 불러오기, 멤버 수락 후 스페이스리스트 및 스페이스 멤버 수 바로 로드

* fix: 스페이스 오너 위임 오류 수정

* fix: 스페이스 오너 위임 후 화면 정보 즉시 갱신

* docs: 리드미 추가, fix: 룸 탈퇴 시 스페이스 오너 검증 추가

* feat: 신청확인에서 수락 시 오류 발생 수정

* fix: 필수 프로필 항목 저장 후 룸 멤버 카드 즉시 갱신

* fix: UserRole ID Long타입으로 변경

* fix: 채팅 및 스페이스, 멤버 목록 영역 스크롤 적용

* FIX : Front_ui

* feat: 테스트데이터

---------


* [Fix] 프론트, 회원 정보 관련 오류 수정 (#72)

* FIX : UI 수정

* feat:

* feat: 멤버 필터링 프론트 추가

* fix: 룸 생성 후 멤버 프로필 카드 작성 화면으로 이동

* FEAT : LocalDateTime -> TIME 으로 변경

* feat: 스페이스 관리에 태그 검색 토글 추가

* feat: 스페이스 정보 수정 커스텀필드 가능하게 변경

* feat: 룸 및 스페이스 멤버 태그 추가

* fix: 모달 리다이렉팅 문제 수정

* feat: 스페이스 관리창에서 위임 이상하게 되던거 수정

* fix: 모달 동작 시 채팅 화면 유지

* feat: 스페이스 참여중/미참여/오너 표시 추가 및 필수항목 표시 추가

* feat: 채팅 순서 및 날짜 추가, 스페이스 열고접기, 필터 위로

* feat: 다른 사람이 보낸 채팅 말풍선 길이 조절

* refactor: 스페이스 탈퇴누르고 확인 누른다음에 채팅 제대로 불러오기, 멤버 수락 후 스페이스리스트 및 스페이스 멤버 수 바로 로드

* fix: 스페이스 오너 위임 오류 수정

* fix: 스페이스 오너 위임 후 화면 정보 즉시 갱신

* docs: 리드미 추가, fix: 룸 탈퇴 시 스페이스 오너 검증 추가

* feat: 신청확인에서 수락 시 오류 발생 수정

* fix: 필수 프로필 항목 저장 후 룸 멤버 카드 즉시 갱신

* fix: UserRole ID Long타입으로 변경

* fix: 채팅 및 스페이스, 멤버 목록 영역 스크롤 적용

* FIX : Front_ui

* feat: 테스트데이터

* fix: 룸 생성 후 멤버 수 오류 수정

* fix: 스페이스 멤버 스크롤 추가

* fix: 회원 정보 수정 오류 수정

* fix: 회원 탈퇴 오류 수정

---------


* FEAT : Add Space, InviteSpace Testcode(#62) (#73)

* DOCS : ADD API명세/기능명세.md db구조.sql

* FIX : MERGE_ERROR

---------

+++ NarangNorang/src/main/java/com/narangnorang/user/service/UserServiceImpl.java
@@ -15,6 +15,10 @@ import com.narangnorang.user.exception.UserException;
 import com.narangnorang.user.exception.errorcode.UserErrorCode;
 import com.narangnorang.user.repository.UserRepository;
 import com.narangnorang.user.repository.UserRoleRepository;
+import com.narangnorang.memberprofilecard.entity.MemberProfileCard;
+import com.narangnorang.memberprofilecard.repository.MemberProfileCardRepository;
+import com.narangnorang.room.repository.RoomRepository;
+import com.narangnorang.space.repository.SpaceRepository;
 
 import jakarta.transaction.Transactional;
 import lombok.RequiredArgsConstructor;
@@ -27,6 +31,9 @@ public class UserServiceImpl implements UserService{
 
 	private final UserRepository userRepository;
 	private final UserRoleRepository userRoleRepository;
+	private final MemberProfileCardRepository memberProfileCardRepository;
+	private final SpaceRepository spaceRepository;
+	private final RoomRepository roomRepository;
 
 	// 사용자 입력 패스워드 (일반 텍스트) 암호화 후 저장
 	// 내맘대로 암호화가 아니라 현재 프로젝트에 설정된 암호화 객체를 이용
@@ -87,10 +94,10 @@ public class UserServiceImpl implements UserService{
 	    String password = resolvePassword(userRequestDto.getPassword(), user.getPassword());
 	    String name = resolveName(userRequestDto.getName(), user.getName());
 
-	    User updatedUser = userRequestDto.toEntity(name, password, user.getUserRoles());
-	    userRepository.save(updatedUser);
+	    user.updateName(name);
+	    user.updatePassword(password);
 
-	    return UserResponseDto.from(updatedUser);
+	    return UserResponseDto.from(user);
 	}
 
 	@Override
@@ -99,6 +106,18 @@ public class UserServiceImpl implements UserService{
 
 	    User user = getUser(userId);
 	    validateOwner(user, email);
+
+	    if (spaceRepository.existsByOwnerId(userId)) {
+	    	throw new UserException(UserErrorCode.SPACE_OWNER_EXISTS);
+	    }
+
+	    if (roomRepository.existsByOwner_Id(userId)) {
+	    	throw new UserException(UserErrorCode.ROOM_OWNER_EXISTS);
+	    }
+
+	    List<MemberProfileCard> profileCards = memberProfileCardRepository.findAllByUserId(userId);
+	    memberProfileCardRepository.deleteAll(profileCards);
+	    memberProfileCardRepository.flush();
 
 	    userRepository.delete(user);
 	}
